address reviewr comments
